Moved test cases to RealmAwareZkClientTestBase. Modified federatedzkclient multi. Fixed minor changes
