+++ zookeeper-api/src/test/java/org/apache/helix/zookeeper/impl/client/TestFederatedZkClient.java
@@ -21,7 +21,6 @@ package org.apache.helix.zookeeper.impl.client;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
@@ -42,9 +41,6 @@ import org.apache.helix.zookeeper.datamodel.serializer.ZNRecordSerializer;
 import org.apache.helix.zookeeper.routing.RoutingDataManager;
 import org.apache.helix.zookeeper.zkclient.IZkStateListener;
 import org.apache.zookeeper.CreateMode;
-import org.apache.zookeeper.Op;
-import org.apache.zookeeper.OpResult;
-import org.apache.zookeeper.ZooDefs;
 import org.apache.zookeeper.Watcher;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -59,9 +55,7 @@ public class TestFederatedZkClient extends RealmAwareZkClientTestBase {
   private static final String TEST_INVALID_PATH = TEST_SHARDING_KEY_PREFIX + "invalid/a/b/c";
   private static final String UNSUPPORTED_OPERATION_MESSAGE =
       "Session-aware operation is not supported by FederatedZkClient.";
-  private static final String PARENT_PATH = ZK_SHARDING_KEY_PREFIX + "/TestFederatedZkClient";
 
-  private RealmAwareZkClient _realmAwareZkClient;
 
   @BeforeClass
   public void beforeClass() throws IOException, InvalidRoutingDataException {
@@ -655,123 +649,11 @@ public class TestFederatedZkClient extends RealmAwareZkClientTestBase {
     }
   }
 
-  /**
-   * Test that zk multi works for op.create.
-   */
-  @Test
-  public void testMultiCreate() {
-    String test_name = "/test_multi_create";
-    _realmAwareZkClient.createPersistent(ZK_SHARDING_KEY_PREFIX);
-
-    //Create Nodes
-    List<Op> ops = Arrays.asList(
-            Op.create(PARENT_PATH, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.create(PARENT_PATH + test_name, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT));
-
-    //Execute transactional support on operations and verify they were run
-    try {
-      List<OpResult> opResults = _realmAwareZkClient.multi(ops);
-      Assert.assertTrue(opResults.get(0) instanceof OpResult.CreateResult);
-      Assert.assertTrue(opResults.get(1) instanceof OpResult.CreateResult);
-      cleanup();
-    } catch (UnsupportedOperationException e) {
-      Assert.fail("Zk Multi Not Supported!");
+  @Override
+  public void testMulti() {
+    super.testMulti();
+    if (!_realmAwareZkClient.exists(ZK_SHARDING_KEY_PREFIX)) {
+      _realmAwareZkClient.createPersistent(ZK_SHARDING_KEY_PREFIX);
     }
-  }
-
-  /**
-   * Multi should be an all or nothing transaction. Creating correct
-   * paths and a singular bad one should all fail.
-   */
-  @Test(dependsOnMethods = "testMultiCreate")
-  public void testMultiFail() {
-    String test_name = "/test_multi_fail";
-    //Create Nodes
-    List<Op> ops = Arrays.asList(
-            Op.create(PARENT_PATH, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.create(PARENT_PATH + test_name, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.create(TEST_INVALID_PATH, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT));
-    try {
-      try {
-        _realmAwareZkClient.multi(ops);
-        Assert.fail("Should have thrown an exception. Cannot run multi on incorrect path.");
-      } catch (Exception e) {
-        boolean pathExists = _realmAwareZkClient.exists(PARENT_PATH);
-        Assert.assertFalse(pathExists, "Path should not have been created.");
-
-        cleanup();
-      }
-    } catch (UnsupportedOperationException e) {
-      Assert.fail("Zk Multi Not Supported!");
-    }
-  }
-
-  /**
-   * Test that zk multi works for delete.
-   */
-  @Test(dependsOnMethods = "testMultiCreate")
-  public void testMultiDelete() {
-    String test_name = "/test_multi_delete";
-    //Create Nodes
-    List<Op> ops = Arrays.asList(
-            Op.create(PARENT_PATH, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.create(PARENT_PATH + test_name, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.delete(PARENT_PATH + test_name, -1));
-
-    try {
-      List<OpResult> opResults = _realmAwareZkClient.multi(ops);
-      Assert.assertTrue(opResults.get(0) instanceof OpResult.CreateResult);
-      Assert.assertTrue(opResults.get(1) instanceof OpResult.CreateResult);
-      Assert.assertTrue(opResults.get(2) instanceof OpResult.DeleteResult);
-
-      cleanup();
-    } catch (UnsupportedOperationException e) {
-      Assert.fail("Zk Multi Not Supported!");
-    }
-  }
-
-  /**
-   * Test that zk multi works for set.
-   */
-  @Test(dependsOnMethods = "testMultiCreate")
-  public void testMultiSet() {
-    String test_name = "/test_multi_set";
-
-    List<Op> ops = Arrays.asList(
-            Op.create(PARENT_PATH, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.create(PARENT_PATH + test_name, new byte[0],
-                    ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT),
-            Op.setData(PARENT_PATH + test_name, new byte[0],
-                    -1));
-
-    try {
-      List<OpResult> opResults = _realmAwareZkClient.multi(ops);
-      Assert.assertTrue(opResults.get(0) instanceof OpResult.CreateResult);
-      Assert.assertTrue(opResults.get(1) instanceof OpResult.CreateResult);
-      Assert.assertTrue(opResults.get(2) instanceof OpResult.SetDataResult);
-
-      cleanup();
-    } catch (UnsupportedOperationException e) {
-      Assert.fail("Zk Multi Not Supported!");
-    }
-  }
-
-  /**
-   * Delete created paths to clean up zk for next multi test case.
-   */
-  public void cleanup() {
-    //Delete Parent path and its children
-    _realmAwareZkClient.deleteRecursively(PARENT_PATH);
-    //Verify path has been deleted
-    boolean pathExists = _realmAwareZkClient.exists(PARENT_PATH);
-    Assert.assertFalse(pathExists, "Parent Path should have been removed.");
   }
 }
